작성자 추가, 로그인 및 수정삭제 권한 체크, 목록에 작성자 이름 나오게

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -12,6 +12,8 @@ import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 
@@ -38,31 +40,81 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
-		Article article = articleService.getArticle(id);
+	public ResultData<Integer> doModify(int id, String title, String body, HttpSession httpSession) {
+		// 로그인 상태 체크
+		boolean isLogined = false;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+		}
+		if (!isLogined) {
+			return ResultData.from("F-A", "로그인하고 이용하세요");
+		}
+		
+		Article article = articleService.getArticle(id); //해당 게시글 가져오기
+		
+		// 게시글 존재여부 체크
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글은 없습니다.", id), id);
 		}
 
+		// 로그인 중인 아이디인지 확인
+		int loginedMemberId = (int) httpSession.getAttribute("loginedMemberId"); // 현재 로그인 중인 번호
+		
+		if(article.getMemberId() != loginedMemberId	) {
+			return ResultData.from("F-B", "해당 게시글에 권한 없음");
+		}
+
+		// 글 수정 작업
 		articleService.modifyArticle(id, title, body);
 		return ResultData.from("S-1", Ut.f("%d번 글을 수정했습니다.", id), id);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(int id, HttpSession httpSession) {
+		// 로그인 상태 체크
+		boolean isLogined = false;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+		}
+		if (!isLogined) {
+			return ResultData.from("F-A", "로그인하고 이용하세요");
+		}
+
+		// 게시글 존재여부 체크
 		Article article = articleService.getArticle(id);
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글은 없습니다.", id), id);
 		}
 
+		// 로그인 중인 아이디인지 확인
+		int loginedMemberId = (int) httpSession.getAttribute("loginedMemberId"); // 현재 로그인 중인 번호
+		
+		if(article.getMemberId() != loginedMemberId	) {
+			return ResultData.from("F-B", "해당 게시글에 권한 없음");
+		}
+		
+		// 글 삭제 작업
 		articleService.deleteArticle(id);
 		return ResultData.from("S-1", Ut.f("%d번 글은 삭제되었습니다.", id), id);
 	}
 
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(String title, String body, HttpSession httpSession) {
+		// 로그인 상태 체크
+		boolean isLogined = false;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+		}
+		if (!isLogined) {
+			return ResultData.from("F-A", "로그인하고 이용하세요");
+		}
+
+		// 제목 내용 작성
 		if (Ut.isNullOrEmpty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
@@ -70,14 +122,16 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body);
+		int loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body, loginedMemberId);
 
 		int id = (int) writeArticleRd.getData1();
 
-		//결과 출력 시 해당 article도 나오게
+		// 결과 출력 시 해당 article도 나오게
 		Article article = articleService.getArticle(id);
 
-		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
+		return ResultData.newData(writeArticleRd, article);
 	}
 
 }
